+++ src/app/home/components/ExperimentPostContainer/ExperimentPostCardListContainer/ExperimentPostCardList/ExperimentPost.css.ts
@@ -1,6 +1,6 @@
 import { style } from '@vanilla-extract/css';
 
-export const postCardListLayout = style({
+export const experimentPostLayout = style({
   display: 'grid',
   gridTemplateColumns: '1fr 1fr 1fr',
   gap: '1.6rem 1.2rem',
