+++ src/app/home/components/ExperimentPostContainer/ExperimentPostCardListContainer/ExperimentPostCardList/ExperimentPostCard/ExperimentPostCard.css.ts
