+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import Banner from './home/components/Banner/Banner';
-import PostContainer from './home/components/PostContainer/PostContainer';
+import ExperimentPostContainer from './home/components/ExperimentPostContainer/ExperimentPostContainer';
 
 import DefaultLayout from '@/components/layout/DefaultLayout/DefaultLayout';
 
@@ -7,7 +7,7 @@ export default function Home() {
   return (
     <DefaultLayout>
       <Banner />
-      <PostContainer />
+      <ExperimentPostContainer />
     </DefaultLayout>
   );
 }
